vbo/requirements.txt
#sudo apt-get install python3.5-dev
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
cachetools==4.1.1
celery==3.1.25
certifi==2020.6.20
chardet==3.0.4
contextlib2==0.6.0.post1
cssselect==1.1.0
cssutils==1.0.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==1.8.19
django-admin-sortable==1.8.3
django-admin-sortable2==0.6.7
django-class-based-auth-views==0.3
django-concurrency==1.2
django-debug-toolbar==1.6
django-dirtyfields==0.1
django-enumfield==1.3b2
django-extensions==2.0.6
django-fsm==2.1.0
django-import-export==0.4.5
django-localflavor-us==1.1
django-mptt==0.7.4
django-redis==4.8.0
django-smart-selects==1.5.1
django-uuidfield==0.5.0
djangorestframework==3.1.1
djoser==0.2.1
et-xmlfile==1.0.1
factory-boy==2.10.0
Faker==4.1.3
FeinCMS==1.12.1
flake8==3.5.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==19.7.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
ipython==5.5.0
ipython-genutils==0.2.0
jdcal==1.4.1
Jinja2==2.8
kombu==3.0.37
lxml==4.5.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mock==2.0.0
mysql-connector-python==8.0.5
mysqlclient==1.4.6
numpy==1.18.5
oauth2client==4.1.3
odfpy==1.4.1
olefile==0.46
opencv-python==3.4.* #sudo apt-get install libsm6 libxrender1 libfontconfig1
openpyxl==2.6.4
parsedatetime==2.6
pbr==5.5.0
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.0b0
Pillow
plivo==0.10.1
premailer==3.7.0
prompt-toolkit==1.0.18
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycurl==7.43.0.6 # sudo apt install libcurl4-openssl-dev libssl-dev
pyflakes==1.6.0
Pygments==2.7.1
PyPDF2==1.26.0
python-dateutil==2.7.0
python-mimeparse==1.6.0
pytz==2020.1
PyYAML==5.3.1
raven==5.9.0
recurrent==0.2.5
redis==3.5.3
reportlab==3.5.50
requests==2.24.0
rsa==4.6
simplegeneric==0.8.1
six==1.15.0
sorl-thumbnail==12.4.1
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.10
validators==0.18.1
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
